Allow Debian package to be built clean-room with pbuilder
The debian control file was missing a dependency on libqt4-dev, so when
pbuilder tried building this package, it would fail for missing
dependencies.

Also, pbuilder by default does not have networking configured (side
effect, builds are reproducible), so we need to pre-download the
wWidgets-1.0.tar.gz file to qt/wwwidgets.